helpers and put in blog edit too
